\name{snpMatrixTranslate-package}
\alias{snpMatrixTranslate-package}
\alias{snpMatrixTranslate}
\docType{package}
\title{
  snpMatrixTranslate
}
\description{
  Write snpMatrix objects to file in other formats.
}
\details{
\tabular{ll}{
Package: \tab snpMatrixTranslate\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-07-06\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Maintainer: Chris Wallace <chris.wallace@cimr.cam.ac.uk>
}
\references{
David Clayton and Hin-Tak Leung (2007). An R package for analysis of whole-genome
association studies. Hum Hered 64:45-51.
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
