% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/write.R
\name{write.impute}
\alias{write.impute}
\title{Write a snpStats object in IMPUTE format}
\usage{
write.impute(X, a1, a2, bp, fileroot, rs.id = NULL, snp.id = NULL)
}
\arguments{
\item{X}{SnpMatrix object}

\item{a1}{vector of first allele at each SNP}

\item{a2}{vector of second allele at each SNP}

\item{bp}{vector of base pair positions for each SNP}

\item{fileroot}{Root of output file names. fileroot.gen and fileroot.sample will be created.}

\item{rs.id}{vector of rs ids (if empty, will set to rownames(X))}

\item{snp.id}{vector of snp ids (if empty, will set to S1...Sn)}
}
\value{
No return value, but has the side effect of writing specified output
files.
}
\description{
see \code{\link{write.simple}} for general information
}
\examples{
data(testdata,package="snpStats")
A.small <- Autosomes[1:6,1:10]
pf <- tempfile()

## write in suitable format for IMPUTE
nsnps <- ncol(A.small)
write.impute(A.small, a1=rep("1",nsnps), a2=rep("2",nsnps), bp=1:nsnps, fileroot=pf)
unlink(pf)
}
\author{
Chris Wallace
}
\keyword{manip}

